pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kaidoku-pdf"
version = "0.1.0"
description = "画像ベースのPDFをOCRし、検索可能なPDF生成やテキスト抽出を行うツールセット"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "PyMuPDF>=1.24.0",
    "pandas>=2.0.0",
]

[project.scripts]
kaidoku-convert = "image_pdf_ocr._cli_convert:main"
kaidoku-extract = "image_pdf_ocr._cli_extract:main"

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-cov",
    "pre-commit",
]

[tool.hatch.build.targets.wheel]
packages = ["image_pdf_ocr"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "UP", "B", "I", "SIM", "RUF"]
ignore = ["RUF001", "RUF002", "RUF003"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=image_pdf_ocr --cov-report=term-missing"
markers = [
    "skipif: conditionally skip tests",
]
